backend/app/requirements.txt
fastapi
uvicorn
rioxarray
rasterio
numpy
matplotlib
pillow
